static_pipeline/requirements.txt
# Basis-Abhängigkeiten für die Static ETL-Pipeline
pandas>=1.5
numpy>=1.23
PyYAML>=6.0
Unidecode>=1.3
matplotlib>=3.7
seaborn>=0.12